Conflicts:
	WebKit/win/WebKit.vcproj/WebKit.sln
	WebKit2/Shared/win/PlatformCertificateInfo.cpp

+++ WebCore/platform/graphics/chromium/DrawingBufferChromium.cpp
@@ -32,6 +32,7 @@
 
 #include "DrawingBuffer.h"
 
+#include "Extensions3DChromium.h"
 #include "GraphicsContext3D.h"
 #include "SharedGraphicsContext3D.h"
 
@@ -48,7 +49,7 @@ struct DrawingBufferInternal {
 #endif
 };
 
-static unsigned generateColorTexture(SharedGraphicsContext3D* context, const IntSize& size)
+static unsigned generateColorTexture(GraphicsContext3D* context, const IntSize& size)
 {
     unsigned offscreenColorTexture = context->createTexture();
     if (!offscreenColorTexture)
@@ -66,13 +67,18 @@ static unsigned generateColorTexture(SharedGraphicsContext3D* context, const Int
 }
 
 
-DrawingBuffer::DrawingBuffer(SharedGraphicsContext3D* context, const IntSize& size, unsigned framebuffer)
+DrawingBuffer::DrawingBuffer(GraphicsContext3D* context, const IntSize& size)
     : m_context(context)
     , m_size(size)
-    , m_framebuffer(framebuffer)
+    , m_fbo(0)
     , m_internal(new DrawingBufferInternal)
 {
-    context->bindFramebuffer(framebuffer);
+    if (!m_context->getExtensions()->supports("GL_CHROMIUM_copy_texture_to_parent_texture")) {
+        m_context.clear();
+        return;
+    }
+    m_fbo = context->createFramebuffer();
+    context->bindFramebuffer(GraphicsContext3D::FRAMEBUFFER, m_fbo);
     m_internal->offscreenColorTexture = generateColorTexture(context, size);
 }
 
@@ -82,14 +88,22 @@ DrawingBuffer::~DrawingBuffer()
     if (m_internal->platformLayer)
         m_internal->platformLayer->setDrawingBuffer(0);
 #endif
-    m_context->bindFramebuffer(m_framebuffer);
+
+    if (!m_context)
+        return;
+        
+    m_context->bindFramebuffer(GraphicsContext3D::FRAMEBUFFER, m_fbo);
     m_context->deleteTexture(m_internal->offscreenColorTexture);
-    m_context->deleteFramebuffer(m_framebuffer);
+
+    clear();
 }
 
 #if USE(ACCELERATED_COMPOSITING)
 void DrawingBuffer::publishToPlatformLayer()
 {
+    if (!m_context)
+        return;
+        
     if (m_callback)
         m_callback->willPublish();
     unsigned parentTexture = m_internal->platformLayer->textureId();
@@ -99,13 +113,16 @@ void DrawingBuffer::publishToPlatformLayer()
     // happens before the compositor draws.  This means we might draw stale frames sometimes.  Ideally this
     // would insert a fence into the child command stream that the compositor could wait for.
     m_context->makeContextCurrent();
-    m_context->copyTextureToParentTextureCHROMIUM(m_internal->offscreenColorTexture, parentTexture);
+    static_cast<Extensions3DChromium*>(m_context->getExtensions())->copyTextureToParentTextureCHROMIUM(m_internal->offscreenColorTexture, parentTexture);
     m_context->flush();
 }
 #endif
 
 void DrawingBuffer::reset(const IntSize& newSize)
 {
+    if (!m_context)
+        return;
+        
     if (m_size == newSize)
         return;
     m_size = newSize;
